requirements.txt
python-telegram-bot==20.7
Flask
requests
gTTS
uuid
